requirements.txt
langchain>=0.1.16
langchain-community>=0.0.13
langchain-core>=0.1.22
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.2
chromadb>=0.4.24
pandas>=2.2.2
streamlit>=1.35.0
sentence-transformers>=2.6.1
huggingface-hub>=0.22.2
python-dotenv>=1.0.0
pydantic>=2.5.2
